+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 0.1.x
 
+### 0.1.1
+
+* Fixed `WordHelper.strip_punctuation` method
+* Fixed empty text and sentiments array handling
+* Added more tests
+
 ### 0.1.0
 
 * Initial release